requirements.txt
opencv-python
requests
Pillow
numpy